// apps/gb-organizer/src/app/core/sidebar.enum.ts
export enum SidebarEnum {
  my_tournaments = 'my_tournaments',
  support = 'support',
  my_plan = 'my_plan',
  merchants = 'Merchants',
  tournament_progress = 'tournament_progress',
  manage_tournament = 'manage_tournament',
  tournament_details = 'tournament_details',
  edit_tournament = 'edit_tournament',
  tournament_logs = 'tournament_logs',
  edit_participants = 'edit_participants',
  manage_winners = 'manage_winners',
  match_list = 'match_list',
  issues_claims = 'issues_claims',
  brackets_view = 'brackets_view',
  tournament_info = 'tournament_info',
  tournament_bracket = 'tournament_bracket',
  participant_settings = 'participant_settings',
  prizes = 'prizes',
  rules = 'rules',
  tournament_setting = 'tournament_setting',
}